Automation Robot
Fire incident is a disaster that can potentially cause the loss of life, property damage and permanent disability to the affected victim. Fire fighters often exposed to higher risks. With the advent of technology, humans are replaced with robots in life-threatening situations. We aim to design a robot capable of detecting and suppressing fires. By designing and implementing a firefighting robot capable of detecting and extinguishing flames, disasters can be avoided with minimal risk to human life. This paper demonstrates the simulation and implementation of an autonomous firefighting robot which can automatically sense the smoke, fire and start to pump water over the flames. Flame and gas sensors were used to detect the fire and smoke. These two sensors can automatically detect fire and smoke & the robot navigates itself to the source of the fire & start extinguishing it by using the fire extinguishing system. This robot also consists with a container on top of the servo motor in order to control the path where water is being sprayed on. Two DC motors were used to control the motor movement while the robot is on operation
mode to extinguish the fire.